# BF/RF: fix command length overflow; cache `OnyoRepo.get_config()`