Fixes in this PR are:

1) don't error when creating a boatload of assets and thus overflowing the maximum command length (#727) 
2) introduce caching for `OnyoRepo.get_config()`
3) cleanup around temp files (dangling file handles, leaked paths, etc)

---
Regarding 1: I introduced a test to overflow with paths (since the overflow can happen with either the message or paths), but left it disabled by default because it takes 30+ minutes to run on my machine. The paths need to be relatively short (beneath the size of the most limiting OS/FS) in order to run reliably. And then in total exceed the maximum argument length of the most capable OS/FS. This results in 150k+ files, and that understandably takes a long time to run.

So the test is there for reference if we are ever uncertain. But is skipped by default.

---
Regarding 2: The motivation is the significant overhead of querying `onyo.assets.name-format`, which is used on a per-asset basis by `set` and `new`.

Initially, I attempted to introduce a per-file file cache at the GitRepo layer, and track the mtime of files that would be queried. However, in the case of a bare `git config` call, the number of locations is quite a few and are modified by environmental variables. These environmental variables would need to be resolved and tracked as well for changes.

CLI users will not have any issues managing the cache (and the new behavior is arguably an improvement, as Onyo's behavior can no longer change mid-flight if the system config changes in parallel). For users of the API, they only encounter issues if they explicitly avoid `OnyoRepo.set_config()`; in which case `clear_cache()` solves their problems.

This speeds up the benchmark considerably:

- `test_cli_onyo_new_fifty[1000] `: 2.09s -> 1.46s (30%)
- `test_cli_onyo_new_fifty[100]`: 1.12s -> 0.42s (63%)
- `test_cli_onyo_set_fifty[1000]`: 2.51s -> 2.31s (8%)
- `test_cli_onyo_set_fifty[100]`: 1.06s -> 0.79s (25%)

Old times:
```
------------------------------------------ benchmark 'test_cli_onyo_new_fifty': 2 tests ------------------------------------------
Name (time in s)                     Min               Max              Mean            StdDev            Median            Rounds
----------------------------------------------------------------------------------------------------------------------------------
test_cli_onyo_new_fifty[1000]     1.8910 (1.74)     2.1347 (1.88)     2.0508 (1.84)     0.0981 (5.01)     2.0877 (1.87)          5
test_cli_onyo_new_fifty[100]      1.0864 (1.0)      1.1338 (1.0)      1.1162 (1.0)      0.0196 (1.0)      1.1177 (1.0)           5
----------------------------------------------------------------------------------------------------------------------------------

------------------------------------------ benchmark 'test_cli_onyo_set_fifty': 2 tests ------------------------------------------
Name (time in s)                     Min               Max              Mean            StdDev            Median            Rounds
----------------------------------------------------------------------------------------------------------------------------------
test_cli_onyo_set_fifty[1000]     2.4798 (2.38)     2.6454 (2.44)     2.5503 (2.40)     0.0700 (4.48)     2.5134 (2.37)          5
test_cli_onyo_set_fifty[100]      1.0427 (1.0)      1.0849 (1.0)      1.0613 (1.0)      0.0156 (1.0)      1.0589 (1.0)           5
----------------------------------------------------------------------------------------------------------------------------------
```

New times:
```
------------------------------------------ benchmark 'test_cli_onyo_new_fifty': 2 tests ------------------------------------------
Name (time in s)                     Min               Max              Mean            StdDev            Median            Rounds
----------------------------------------------------------------------------------------------------------------------------------
test_cli_onyo_new_fifty[1000]     1.4381 (3.47)     1.6802 (3.61)     1.5081 (3.48)     0.0999 (4.20)     1.4568 (3.49)          5
test_cli_onyo_new_fifty[100]      0.4146 (1.0)      0.4654 (1.0)      0.4329 (1.0)      0.0238 (1.0)      0.4178 (1.0)           5
----------------------------------------------------------------------------------------------------------------------------------

------------------------------------------ benchmark 'test_cli_onyo_set_fifty': 2 tests ------------------------------------------
Name (time in s)                     Min               Max              Mean            StdDev            Median            Rounds
----------------------------------------------------------------------------------------------------------------------------------
test_cli_onyo_set_fifty[1000]     2.2197 (2.83)     2.3716 (2.99)     2.2920 (2.90)     0.0629 (15.11)    2.3075 (2.92)          5
test_cli_onyo_set_fifty[100]      0.7846 (1.0)      0.7940 (1.0)      0.7897 (1.0)      0.0042 (1.0)      0.7911 (1.0)           5
----------------------------------------------------------------------------------------------------------------------------------
```

---
fix #727